matched

+++ host/repositories/project/project.js
@@ -1,8 +1,10 @@
 import Group from "../../models/groupModel.js";
+import Matched from "../../models/matchedModel.js";
 import MentorCategory from "../../models/mentorCategory.js";
 import ProjectCategory from "../../models/projectCategory.js";
 import Project from "../../models/projectModel.js";
 import TemporaryMatching from "../../models/temporaryMatching.js";
+import User from "../../models/userModel.js";
 
 const createProject = async (projectData) => {
   try {
@@ -15,20 +17,27 @@ const createProject = async (projectData) => {
 const updateProject = async (id, project) => {
   try {
     const result = await Project.findByIdAndUpdate(id, project);
-    const listProjectCategories = await ProjectCategory.find({
-      projectId: result._id,
-    });
-    const listCategoryId = listProjectCategories.map((c) => c.categoryId);
-    const mentorCategories = await MentorCategory.find({
-      categoryId: { $in: listCategoryId },
-    });
     const group = await Group.findOne({ projectId: result._id });
-    const temporaryMatching = mentorCategories.map((c) => ({
-      groupId: group._id,
-      mentorId: c.userId,
-    }));
-    await TemporaryMatching.deleteMany({ groupId: group._id });
-    await TemporaryMatching.insertMany(temporaryMatching);
+    const matched = Matched.findOne({ projectId: result._id });
+    if (!matched) {
+      const listProjectCategories = await ProjectCategory.find({
+        projectId: result._id,
+      });
+      const listCategoryId = listProjectCategories.map((c) => c.categoryId);
+      const mentorCategories = await MentorCategory.find({
+        categoryId: { $in: listCategoryId },
+      });
+      const user = await User.findById(mentorCategories[0].userId);
+      const matchedQtt = await Matched.countDocuments({ mentorId: user._id });
+      const temporaryMatching = mentorCategories.map((c) => ({
+        groupId: group._id,
+        mentorId: c.userId,
+      }));
+      if (user.menteeCount > matchedQtt) {
+        await TemporaryMatching.deleteMany({ groupId: group._id });
+        await TemporaryMatching.insertMany(temporaryMatching);
+      }
+    }
     return result;
   } catch (error) {
     throw new Error(error.message);
